add the webpack + refcator code
