+++ src/app.js
@@ -1,0 +1,24 @@
+function deleteOption(button, book, card) {
+  button.addEventListener('click', () => {
+    card.parentNode.removeChild(card);
+    if (myLibrary.indexOf(book) !== -1) {
+      myLibrary.splice(myLibrary.indexOf(book), 1);
+    }
+  });
+}
+ 
+function toggleStatus(button, book, card) {
+  button.addEventListener('click', () => {
+    const status = card.querySelector('.read-status');
+    if (book.read === 'Read') {
+      status.textContent = 'Not Read';
+      book.read = 'Not Read';
+    } else {
+      status.textContent = 'Read';
+      book.read = 'Read';
+    }
+  });
+}
+
+
+export { deleteOption, toggleStatus };
